requirements.txt
matplotlib
nnfs
numpy
# opencv-python
pillow